Everywhere: Initial commit
This patch initializes the `tokio-unix-tcp` library, wrapping tokio
networking types for tcp and unix sockets. It includes utility types
and functions for addresses, binding and connecting and should work
as a drop in replacement for types in a TCP tokio stack.

The library supports serializing and deserializing address wrappers
with serde and compiles unix functionality to no ops on non unix
system.