Serde: Use default derive Serialize and Deserialize for SocketAddrs
This patch removes the custom serialize and deserialize impls for
`SocketAddr` and `NamedSocketAddr`.

Previously, these implementations were incompatible and serialized
output was not able to be deserialized again. This was due to the
deserialization implementation taking a string as input to be used
when reading configuration files.

Now both enums use the default derive implementations and include
additional helper functions to deserialize string representations
to `SocketAddr`, `NamedSocketAddr`, `Option<SocketAddr>` or
`Option<NamedSocketAddr>`.